First compilation of the Python-compatible version.
Managed to import the rovib code as a Python library. This was possible by referencing all the Fortran libraries used in rovib.f90, removing a layer of indentations at the Fortran code and labeling some variables with intents for f2py to recognize them as inputs and outputs.
TODO: Fix array compatibility and make the main program code a callable subroutine.